% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlFunctions.R
\name{draftResult}
\alias{draftResult}
\title{Looks into the DB for what happened that day for draft}
\usage{
draftResult(today, teamID)
}
\arguments{
\item{today}{todays date}
}
\description{
Looks into the DB for what happened that day for draft
}
