% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlFunctions.R
\name{teamResult}
\alias{teamResult}
\title{Looks into the DB for what happened that day}
\usage{
teamResult(today, teamID)
}
\arguments{
\item{today}{todays date}
}
\description{
Looks into the DB for what happened that day
}
