% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlFunctions.R
\name{tradeResult}
\alias{tradeResult}
\title{Looks into the DB for what happened that day for trades}
\usage{
tradeResult(today)
}
\arguments{
\item{today}{todays date}
}
\description{
Looks into the DB for what happened that day for trades
}
