% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlFunctions.R
\name{tradeResultAll}
\alias{tradeResultAll}
\title{Looks into the DB for what trades happened that day}
\usage{
tradeResultAll(today)
}
\arguments{
\item{today}{todays date}
}
\description{
Looks into the DB for what trades happened that day
}
